refactor - working
