+++ code/spikes/publish.001/src/types.mts
@@ -1,5 +1,4 @@
 export * from './Pkg.Content.Bundle/types.mjs';
 export * from './Pkg.Content.Log/types.mjs';
 export * from './Markdown.File/types.mjs';
-
-export * from './ui.logic/state/types.mjs';
+export * from './ui.logic/types.mjs';
