+++ code/spikes/publish.001/src/ui.logic/common.mts
@@ -1,1 +1,1 @@
-export * from '../ui/common/index.mjs';
+export * from './common/index.mjs';
