+++ code/spikes/publish.001/src/ui.logic/common/types.mts
@@ -1,0 +1,10 @@
+/**
+ * @system
+ */
+export type { MdastRoot, MdastHeading, MdastCode, MdastText } from 'sys.text/src/types.mjs';
+export type { Text, MarkdownProcessor } from 'sys.text/src/types.mjs';
+
+/**
+ * @local
+ */
+export * from '../../types.mjs';
