--- code/spikes/publish.001/src/ui.logic/state/index.mts
@@ -1,2 +1,0 @@
-export * from './State.mjs';
-export * from './QueryString.mjs';
