+++ code/spikes/publish.001/src/ui/Markdown/Markdown.tsx
@@ -1,8 +1,9 @@
 import React, { useEffect, useState } from 'react';
 
-import { css, t, State } from '../common.mjs';
+import { css, t } from '../common.mjs';
 import { MarkdownEditor } from './Markdown.Editor';
 import { MarkdownOutline } from './Markdown.Outline';
+import { State } from '../../ui.logic/index.mjs';
 
 export type MarkdownProps = {
   location: string;
