--- code/spikes/publish.001/src/ui/common/colors.mts
@@ -1,8 +1,0 @@
-export const COLORS = {
-  BLACK: '#000',
-  WHITE: '#fff',
-  DARK: '#293042', // Inky blue/black.
-  CYAN: '#00C2FF',
-  MAGENTA: '#FE0064',
-  BLUE: '#4D7EF7',
-};
