--- code/spikes/publish.001/src/ui/common/libs.mts
@@ -1,12 +1,0 @@
-/**
- * @external
- */
-import { groupBy, uniq } from 'ramda';
-export const R = { groupBy, uniq };
-
-/**
- * @system
- */
-export { rx, slug } from 'sys.util';
-export { FC } from 'sys.util.react';
-export { css, Color, Style } from 'sys.util.css';
